Slaves market racer

+++ database/migrations/2024_09_21_190530_create_transactions_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('transactions', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('wallet_id')->nullable();
-            $table->float('value');
+            $table->decimal('value', 15, 2);
             $table->timestamps();
 
             $table->foreign('wallet_id')->references('id')->on('wallets');
